<fix>配置文件接口可直接接收json字符串形式的配置信息 <fix>优化字符画Python接口，可直接print量子程序 <fix>量子线路适配芯片拓扑结构python接口封装 <fix>添加Grover算法python接口 <fix>添加获取量子线路所用到的qubit的python接口 去掉不必要的log记录，仅保留报错log <fix>添加线路适配异常处理 <fix>修改windows控制台字符画乱码问题 <feat> MPS 振幅模拟功能 <fix> 修改FindQPANDA.cmake <fix>优化了测量部分 <fix> 添加元数据头文件 <fix>applications 中 HHL 实例修改 <fix>元数据校验不通过时，抛出异常 <fix>虚拟z门功能添加U3门优化 <fix>添加多控门分解独立接口 <fix>修改测试用例中命名不规范问题 <fix>取消芯片拓扑结构孤立点判断 <fix> 修改QProgCheck中控制逻辑门的信息统计 <fix>虚拟z门转换添加串扰补偿功能 <fix>量子芯片适配功能中，在基础双门转换前，进行线路优化 <fix> 修改runwithconfiguration bug <fix>修改线路替换时，qubit内存错误问题 <fix>删除test分支虚拟z门相关代码 <fix>测量同一比特以及测试结果存在同一经典比特情况的处理 <fix> 头文件添加到core.h中 <fix> 修改U4、CU、QDouble接口及pyqpanda编译错误 <fix>更新originir支持PI，自然记数以及RPhi门 <fix> 修改python中的bug <fix>修改部分振幅的一个问题 <fix>修改量子云 <fix>添加MPS噪声 <fix>QCloudMachine更新 <fix>解决linux编译错误问题 <fix>优化了接口的使用 <fix>private分支代码合并 <fix>修改了QCloudMachine的错误状态提示 <fix>请求接口加入默认参数taskname <fix>删除部分注释 <fix> 映射支持 ISWAPTheta门 <fix>解决mac下编译错误 <fix>解决mac编译错误 <fix>修改QPanda发版库中的bug <fix>修改受控单门无法在多控门分解接口分解的问题 <fix> update qasm to qprog CR

+++ include/Core/Utilities/QPandaNamespace.h
@@ -28,7 +28,7 @@
 * @def qstate_type
 * @brief QPanda2 quantum state data type
 */
-typedef float qstate_type;
+typedef double qstate_type;
 
 /**
 * @def qcomplex_t
